refactor: shuffle code to their proper place
